Fixed spatial Hadamard tests: property access, pipe verification, correlation surfaces, ungraphlike errors
